User commit

+++ src/main/java/com/example/demo/repository/IdDetailsRepo.java
@@ -3,8 +3,8 @@ package com.example.demo.repository;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import com.example.demo.entity.IdDetailsRepo;
-import com.example.demo.entity.ValidationEntity;
+import com.example.demo.entity.IdDetails;
 @Repository
-public interface IdDetailsRepo extends JpaRepository<IdDetails,Long>{
+public interface IdDetailsRepo extends JpaRepository<IdDetails,Integer>{
 
 }
